+++ internal/models/token.go
@@ -1,0 +1,7 @@
+package models
+
+var RefreshTokenName string = "rtokens"
+
+type RefreshToken struct {
+	Token string `db:"token"`
+}