+++ migrations/sqlite3/20220222194318-token.sql
@@ -1,0 +1,7 @@
+
+-- +migrate Up
+CREATE TABLE rtokens (
+    token VARCHAR(511) NOT NULL UNIQUE PRIMARY KEY
+);
+-- +migrate Down
+DROP TABLE rtokens;